Move state code into CaseloadItemViewModel and dispose records before clearing
